[android] Fixed energy being double.

+++ android/app/src/main/java/rocks/crownstone/consumerapp/BluenetBridge.java
@@ -2639,7 +2639,7 @@ public class BluenetBridge extends ReactContextBaseJavaModule implements Interva
 			serviceDataMap.putDouble("powerFactor", 1);
 			serviceDataMap.putDouble("powerUsageReal", serviceData.getPowerUsage() / 1000.0);
 			serviceDataMap.putDouble("powerUsageApparent", serviceData.getPowerUsage() / 1000.0);
-			serviceDataMap.putInt("accumulatedEnergy", serviceData.getAccumulatedEnergy());
+			serviceDataMap.putDouble("accumulatedEnergy", serviceData.getAccumulatedEnergy());
 			serviceDataMap.putBoolean("newDataAvailable", CrownstoneServiceData.isNewData(eventBitmask));
 			serviceDataMap.putBoolean("stateOfExternalCrownstone", CrownstoneServiceData.isExternalData(eventBitmask));
 			serviceDataMap.putBoolean("setupMode", serviceData.isSetupMode());
